% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CredibleInt.R
\name{CredibleInt}
\alias{CredibleInt}
\title{Construct Credible Interval for survival curve}
\usage{
CredibleInt(PosteriorDraws, CredLevel = 0.95)
}
\arguments{
\item{PosteriorDraws}{Matrix of draws from the posterior distribution
of the survival curve. Columns correspond to all draws at fixed time.
Rows correspond to single draw of survival curve over all observed
time points.}

\item{CredLevel}{The Credible Level of the desired credible interval.}
}
\description{
Function which uses draws from the posterior distribution at each
time point to construct a credible interval with set credible level
}
