% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatherResults.R
\name{GatherResults}
\alias{GatherResults}
\title{Read in SurvBART output data from directory}
\usage{
GatherResults(file.dir = NULL)
}
\arguments{
\item{file.dir}{Character object specying the location of the saved data}
}
\description{
Reads in data from specified directory
}
