% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KMQuant.R
\name{KMQuant}
\alias{KMQuant}
\title{Estimate Quantiles using Kaplan-Meier Product Limit Estimator}
\usage{
KMQuant(Times, Event, Quantiles, Predictors = 1, ConfLevel)
}
\arguments{
\item{Times}{Observed times of event or right-censoring}

\item{Event}{Indicator of whether a unit is observed (1) or censored(0)}

\item{Quantiles}{The quantiles we wish to estimate}

\item{Predictors}{Categorical predictor variable}

\item{ConfLevel}{The confidence level for the quantile confidence
interval}
}
\description{
Estimates quantiles of a survival function using the Kaplan-Meier Product
Limit Estimator. Returns a matrix of point estimates of the quantiles
along with confidence intervals.
}
