% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObtainPost.R
\name{ObtainPost}
\alias{ObtainPost}
\title{Obtain Posterior Draws}
\usage{
ObtainPost(TrainX = NULL, Times, Event, TestX = NULL, parallel = F,
  NumCores = 2, seed = 807)
}
\arguments{
\item{TrainX}{Explanatory variables for training (in sample) data. 
Must be a matrix with rows corresponding to observations and columns to variables}

\item{Times}{The time of event or right-censoring}

\item{Event}{The event indicator: 1 is and event while 0 is censored}

\item{TestX}{Explanatory variables for test (out of sample) data. Must be a matrix and 
have the same structure as TrainX}

\item{parallel}{Whether to obtain posterior draws across multiple cores.}

\item{NumCores}{Number of cores to run on, default is 2}

\item{seed}{Random seed, only used for parallelization}
}
\description{
Function to obtain draws from the posterior distribution of the survival curve.
}
