% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParallelUnix.R
\name{ParallelUnix}
\alias{ParallelUnix}
\title{Parallelize posterior draws for unix machine}
\usage{
ParallelUnix(TrainX, Times, Event, TestX, NumCores, seed)
}
\arguments{
\item{TrainX}{Explanatory variables for training (in sample) data. 
Must be a matrix with rows corresponding to observations and columns to variables}

\item{Times}{The time of event or right-censoring}

\item{Event}{The event indicator: 1 is and event while 0 is censored}

\item{TestX}{Explanatory variables for test (out of sample) data. Must be a matrix and 
have the same structure as TrainX}

\item{NumCores}{Number of cores to run on, default is 2}

\item{seed}{Random seed, only used for parallelization}
}
\description{
Function which runs in parallel to obtain posterior draws of the 
survival function on a unix based machine
}
