% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParallelWindows.R
\name{ParallelWindows}
\alias{ParallelWindows}
\title{Obtain Posterior Draws on a non unix machine}
\usage{
ParallelWindows(TrainX, Times, Event, TestX, NumCores)
}
\arguments{
\item{TrainX}{Explanatory variables for training (in sample) data. 
Must be a matrix with rows corresponding to observations and columns to variables}

\item{Times}{The time of event or right-censoring}

\item{Event}{The event indicator: 1 is and event while 0 is censored}

\item{TestX}{Explanatory variables for test (out of sample) data. Must be a matrix and 
have the same structure as TrainX}

\item{NumCores}{Number of cores to run on, default is 2}
}
\description{
Function which runs in parallel to obtain posterior draws of the survival function 
on a non-unix based machine.
}
