% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSurv.R
\name{PlotSurv}
\alias{PlotSurv}
\title{Plot BART estimated Survival Curves}
\usage{
PlotSurv(PosteriorDraws, UniqueTimes, CredLevel = 0.95)
}
\arguments{
\item{PosteriorDraws}{Matrix of draws from the posterior distribution
of the survival function}

\item{UniqueTimes}{The unique times of observation or censoring}

\item{CredLevel}{The desired credible level}
}
\description{
This function allows the user to plot the estimated survival curve
along with its credible interval.
}
