% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuantAll.R
\name{QuantAll}
\alias{QuantAll}
\title{Estimate Survival Quantiles using BART, KM, and Weibull Regression}
\usage{
QuantAll(PosteriorDraws, Times, Event, Predictors = 1, Quantiles,
  ConfLevel)
}
\arguments{
\item{PosteriorDraws}{Draws from the posterior distribution of the
survival function}

\item{Times}{Observed time to events or right-censoring}

\item{Event}{Vector of event status: 1 = Observed, 0 = Censored}

\item{Predictors}{Desired model predictors}

\item{Quantiles}{Quantiles to be estimated}

\item{ConfLevel}{Desired confidence level to construct confidence
intervals}
}
\description{
Estimate Survival Quantiles and associated confidence bounds using
BART, KM, and Weibull Regression.
}
