% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuantEst.R
\name{QuantEst}
\alias{QuantEst}
\title{Estimate quantiles of survival function}
\usage{
QuantEst(PosteriorDraws, Times, Quantiles, CredLevel)
}
\arguments{
\item{PosteriorDraws}{Matrix of random draws from the posterior distribution
of the survival curve.}

\item{Times}{Vector of unique time points}

\item{Quantiles}{Quantiles to be estimated}

\item{CredLevel}{Credible level for upper and lower bounds of quantile estimate}
}
\description{
Estimates the quantiles of the survival function along with credible bounds
based upon draws from the posterior distribution
}
