% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateData.R
\name{SimulateData}
\alias{SimulateData}
\title{Simulate Data}
\usage{
SimulateData(n, censoring.rate, TwoSample = F)
}
\arguments{
\item{n}{Desired sample size}

\item{censoring.rate}{Desired censoring rate, either 0.2 or 0.5}

\item{TwoSample}{Do you want to simulate under the two-sample settings
described in Sparapani (2016)?}
}
\description{
A function to simulate data as specified in Sparparani et al. (2016)
}
\examples{
dat = SimulateData(n = 100, censoring.rate = 0.2)
head(dat)
}
\keyword{simulate}
\keyword{weibull}
