% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{Simulation}
\alias{Simulation}
\title{Perform one iteration of simulation}
\usage{
Simulation(OneSample = T, n, CensoringRate)
}
\arguments{
\item{OneSample}{Perform simulation described in "One-Sample Simulation" section}

\item{n}{Sample Size}

\item{CensoringRate}{Desired Censoring Rate, either 0.2 or 0.5}
}
\description{
Performs one iteration of the simulation described in Sparapani et al.
}
