% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvEst.R
\name{SurvEst}
\alias{SurvEst}
\title{Estimate the Survival Curve}
\usage{
SurvEst(PosteriorDraws)
}
\arguments{
\item{PosteriorDraws}{Matrix of random draws from the posterior distribution
of the survival curve.}
}
\description{
Return the predicted survival probability at each unique time point
}
