% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeibQuant.R
\name{WeibQuant}
\alias{WeibQuant}
\title{Obtain estimated Quantiles from Weibull regression}
\usage{
WeibQuant(Times, Event, Predictors, Quantiles, ConfLevel)
}
\arguments{
\item{Times}{Observed time to event or right-censoring}

\item{Event}{Indicator of event observed (1) or censored(0)}

\item{Predictors}{The desired predictors of survival}

\item{Quantiles}{Quantiles we wish to estimate}

\item{ConfLevel}{The desired confidence level on which to construct
confidence intervals for the quantiles}
}
\description{
Gives estimated quantiles and associated Wald confidence intervals
from a Weibull regression model.
}
