% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyzeResults.R
\name{AnalyzeResults}
\alias{AnalyzeResults}
\title{Analyse results}
\usage{
AnalyzeResults(Results)
}
\arguments{
\item{Results}{A list of results from multiple runs of Simulation()}
}
\description{
This function analyses results from the Simulation() function. Given a list of results
from many runs of Simulation(), this function outputs a matrix of bias, RMSE, and coverage
probabilities
}
