% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyzeResults.R
\name{AnalyzeResults}
\alias{AnalyzeResults}
\title{Analyse results}
\usage{
AnalyzeResults(Results, ReturnOOR = FALSE)
}
\arguments{
\item{Results}{A list of results from multiple runs of Simulation()}

\item{ReturnORR}{Logical value. Do you want to know how many quantile estimates
were inestimable because the quantile was outside of the range of the data?}
}
\description{
This function analyses results from the Simulation() function. Given a list of results
from many runs of Simulation(), this function outputs a matrix of bias, RMSE, and coverage
probabilities
}
