% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{Simulation}
\alias{Simulation}
\title{Perform one iteration of simulation}
\usage{
Simulation(OneSample = T, n, CensoringRate, Parallel = F,
  NumCores = 2, seed, save.output = F, file.path = NULL)
}
\arguments{
\item{OneSample}{Perform simulation described in "One-Sample Simulation" section}

\item{n}{Sample Size}

\item{CensoringRate}{Desired Censoring Rate, either 0.2 or 0.5}

\item{Parallel}{Run the function in parallel}

\item{NumCores}{If parallel, number of cores to use}

\item{seed}{Set random seed number}

\item{save.output}{Do you want to save results?}

\item{file.path}{Where to save}
}
\description{
Performs one iteration of the simulation described in Sparapani et al.
}
